Don't call daemon() and setup our signal handlers until after we check and do the unregister/reregister work.
Don't call syslog in the unregister/reregister code as we haven't called
openlog() yet.

Be a more conservative about accepting errno values from socket(2),
only EPROTONOSUPPORT means that the kernel isn't supporting it
something like INET6.  The other possible errnos would be returned
if there was a mistake in the socket(2) call so remove them from the
list of "acceptable" return values.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@74800 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f