(forced commit for more info on the previous change)
According to RFC 1112, which deals with multicasting, an application must
specify the interface on which it wished to send a multicast packet. In the
absence of such an interface, a default is selected.

Previous behavior in ip_output.c erroneously checked for a route to the
destination multicast address, rejecting the packet if none existed.

Applications got around this with a FAQ recommending a 224.0.0.0/29 route
to loopback. This worked because the multicast code in ip_output.c discards
route information if an interface was selected.

The previous commit skips the route check in the case where a multicast packet
is being send to an application-defined interface.

The only change in behavior, if no bugs were introduced, is the lack of
rejection of a multicast packet for which no route exists sent to an
application-defined interface.

MFC after:	2 weeks


git-svn-id: svn+ssh://svn.freebsd.org/base/head@79836 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f