Work-in-progress commit syncing up pmap cleanups that I have been working on for a while: - fine grained TLB shootdown for SMP on i386 - ranged TLB shootdowns.. eg: specify a range of pages to shoot down with   a single IPI, since the IPI is very expensive.  Adjust some callers   that used to trigger this inside tight loops to do a ranged shootdown   at the end instead. - PG_G support for SMP on i386 (options ENABLE_PG_G) - defer PG_G activation till after we decide what we are going to do with   PSE and the 4MB pages at the start of the kernel.  This should solve   some rumored strangeness about stale PG_G entries getting stuck   underneath the 4MB pages. - add some instrumentation for the fine TLB shootdown - convert some asm instruction wrappers from functions to inlines.  gcc   seems to do a fair bit better with this. - [temporarily!] pessimize the tlb shootdown IPI handlers.  I will fix   this again shortly.
This has been working fairly well for me for a while, but I have tweaked
it again prior to commit since my last major testing round.  The only
outstanding problem that I know of is PG_G related, which is why there
is an option for it (not on by default for SMP).  I have seen a world
speedups by a few percent (as much as 4 or 5% in one case) but I have
*not* accurately measured this - I am a bit sceptical of these numbers.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@91260 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f