Move all remaining tests except for: 	make(1):	Does not work like the other tests.  Its Makefile is 			self-testing. 	m4(1):		It uses complex voodo to test GNU m4(1) features.
To the new framework.  I had worried about passing the binary data that
uudecode(1)'s test passes to diff(1) might give a user something nasty,
but this is unlikely to happen as even with an unmodified old nasty
diff(1) which doesn't recognise many binary files, these binary files
are recognised.  Using $DIFF instead of `diff' in the library and making
it possible to override this with `cmp -s' might be nice some day, but
as of this second, there's no immediate need.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@98759 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f