Keep a copy of the credential used to mount filesystems around so we can check and use it later on.
Change the pieces of code which relied on mount->mnt_stat.f_owner
to check which user mounted the filesystem.

This became needed as the EA code needs to be able to allocate
blocks for "system" EA users like ACLs.

There seems to be some half-baked (probably only quarter- actually)
notion that the superuser for a given filesystem is the user who
mounted it, but this has far from been carried through.  It is
unclear if it should be.

Sponsored by: DARPA & NAI Labs.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@102088 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f