struct tm.tm_year is listed as 'years since 1900', and is signed.  On 64 bit systems, years roughly -2^31 through 2^31 can be represented in time_t without any trouble.  32 bit time_t systems only range from roughly 1902 through 2038.  As a consequence, none of the date munging code for all the various calendar tweaks before then is present.  There are other problems including the fact that there was no 'year zero' and so on.  So rather than get excited about trying to figure out when the calendar jumped by two weeks etc, simply disallow negative (ie: prior to 1900) years.
This happens to have an important side effect.  If you bzero a 'struct
tm', it corresponds to 'Jan 0, 1900, 00:00 GMT'.  This happens to be
representable (after canonification) in 64 bit time_t space.  Zero tm
structs are generally an error and mktime normally returns -1 for them.
Interestingly, it tries to canonify the 'jan 0' to 'dec 31, 1899', ie:
year -1.  This conveniently trips the negative year test above, which
means we can trivially detect the null 'tm' struct.

This actually tripped up code at work. :-/  (Don't ask)


git-svn-id: svn+ssh://svn.freebsd.org/base/head@134231 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f